removed the old TripManagement file
